requirements.txt
Django==2.0.1
django-wysiwyg==0.8.0
django-tastypie==0.14.0